added supertest and one GET test
